### 2024-12-04 #7567 @editorialbot
[PRE REVIEW]: SwiftPol: A Python package for building and parameterizing in silico polymer systems
**Submitting author:** @hannaomi (<a href="http://orcid.org/0009-0002-3298-0309">Hannah Turney</a>)
**Repository:** https://github.com/matta-research-group/SwiftPol
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.1.0
