### 2024-12-04 2c12d24
added option to not overwrite partial charges
