+++ swiftpol/parameterize.py
@@ -62,35 +62,49 @@ def charge_polymer(polymer, charge_scheme):
         raise AttributeError("This function takes either 'AM1_BCC', 'NAGL', or 'espaloma' as charge_scheme input")
     
 
-def charge_openff_polymer(openff_chain, charge_scheme):
+def charge_openff_polymer(openff_chain, charge_scheme, overwrite=True):
     """
-    Calculate and return the partial charges of a polymer chain based on the specified charge scheme.
+    Assign partial charges to a polymer chain based on the specified charge scheme.
+
+    This function assigns partial charges to a polymer chain using one of the following charge schemes:
+    'AM1_BCC', 'NAGL', or 'espaloma'. It can optionally overwrite existing charges.
 
     Parameters
     ----------
-    polymer : rdkit.Chem.rdchem.Mol
-        A polymer chain for which the charges are to be calculated.
+    openff_chain : openff.toolkit.topology.Molecule
+        The original polymer chain to which partial charges will be assigned.
     charge_scheme : str
-        A string that specifies the charge scheme to be used. It can be either 'AM1_BCC', 'espaloma', or 'NAGL'.
+        The charge scheme to use for assigning partial charges. Options are 'AM1_BCC', 'NAGL', or 'espaloma'.
+    overwrite : bool, optional
+        Whether to overwrite existing partial charges. Default is True.
 
     Returns
     -------
-    numpy.ndarray
-        The partial charges of the polymer chain according to the specified charge scheme.
+    openff.toolkit.topology.molecule.PartialChargeArray
+        The partial charges assigned to the polymer chain.
 
     Raises
     ------
+    ModuleNotFoundError
+        If the installed version of openff-toolkit is below what is required to use NAGL.
+    ImportError
+        If the required package for the specified charge scheme is not installed.
     AttributeError
         If the charge_scheme input is not 'AM1_BCC', 'NAGL', or 'espaloma'.
     """
     # Function implementation here
     from openff import toolkit
     from openff.toolkit.topology import Molecule
+    chain_copy = openff_chain
     if charge_scheme == 'AM1_BCC':
-        if openff_chain.conformers is None:
-            openff_chain.generate_conformers()
-        openff_chain.assign_partial_charges("am1bcc")
-        return openff_chain.partial_charges
+        if chain_copy.conformers is None:
+            chain_copy.generate_conformers()
+        if overwrite:
+            openff_chain.assign_partial_charges("am1bcc")
+            return openff_chain.partial_charges
+        else:
+            chain_copy.assign_partial_charges("am1bcc")
+            return chain_copy.partial_charges
     elif charge_scheme == 'NAGL' and toolkit.__version__ < '0.16.0':
         raise ModuleNotFoundError("Installed version of openff-toolkit is below what is required to use NAGL. Please update to v.0.16.0")
     elif charge_scheme == 'NAGL' and toolkit.__version__ >= '0.16.0':
@@ -99,16 +113,24 @@ def charge_openff_polymer(openff_chain, charge_scheme):
         except:
             raise ImportError("The package openff-nagl is not installed. You will not be able to use NAGL.")
         ntkw = NAGLToolkitWrapper()
-        ntkw.assign_partial_charges(openff_chain, "openff-gnn-am1bcc-0.1.0-rc.2.pt")
-        return openff_chain.partial_charges
+        if overwrite:
+            ntkw.assign_partial_charges(openff_chain, "openff-gnn-am1bcc-0.1.0-rc.2.pt")
+            return openff_chain.partial_charges
+        else:
+            ntkw.assign_partial_charges(chain_copy, "openff-gnn-am1bcc-0.1.0-rc.2.pt")
+            return chain_copy.partial_charges
[…]
