+++ swiftpol/_version.py
@@ -1,1 +1,1 @@
-__version__ = "1.0.0+229.g8daf410.dirty"
+__version__ = "1.0.0+235.g2c12d24.dirty"
