### 2024-12-09 3e7740f
add names to residual monomer openff objects
