### 2024-12-10 60cecf8
added preliminary tacticity support
