### 2024-12-10 af2014c
Updated docstring for build.py tacticity
