### 2024-12-10 898601d
2nd option for blockiness calculation method calculation
