+++ swiftpol/build.py
@@ -246,7 +246,7 @@ def PDI(chains):
 
 
 
-def blockiness_gen(sequence):
+def blockiness_gen(sequence, wrt='A'):
     """
     Calculate the blockiness and average block length of a co-polymer sequence.
 
@@ -270,26 +270,53 @@ def blockiness_gen(sequence):
     -----
     If the sequence does not contain both 'A' and 'B', the function returns a string indicating that the molecule is not a co-polymer.
     """
+    if wrt == 'A':
+        assert 'A' in sequence, "Sequence does not contain monomer A, cannot calculate blockiness with reference to A"
+        if 'A' in sequence and 'B' in sequence: #Check if sequence is a co-polymer
+            AB = sequence.count('AB')
+            BB = sequence.count('BB')
+            BA = sequence.count('BA')
+            AA = sequence.count('AA')
+            if 'BA' in sequence:
+                blockiness = AA/BA
+            else:
+                blockiness = AA/AB
+            #Calculate block length B
+            block_list_B = [x for x in sequence.split('A') if x!='']
+            block_length_B = mean([len(b) for b in block_list_B])
+            #Calculate block length A
+            block_list_A = [x for x in sequence.split('B') if x!='']
+            block_length_A = mean([len(b) for b in block_list_A])
+            return blockiness, block_length_B, block_length_A
 
-    if 'A' in sequence and 'B' in sequence: #Check if sequence is a co-polymer
-        AB = sequence.count('AB')
-        BB = sequence.count('BB')
-        BA = sequence.count('BA')
-        AA = sequence.count('AA')
-        if 'BA' in sequence:
-            blockiness = BB/BA
         else:
-            blockiness = BB/AB
-        #Calculate block length B
-        block_list_B = [x for x in sequence.split('A') if x!='']
-        block_length_B = mean([len(b) for b in block_list_B])
-        #Calculate block length A
-        block_list_A = [x for x in sequence.split('B') if x!='']
-        block_length_A = mean([len(b) for b in block_list_A])
-        return blockiness, block_length_B, block_length_A
+            return 'Molecule is not a co-polymer, no blockiness calculation performed', 0, len(sequence)
+    
+    elif wrt == 'B':
+        assert 'B' in sequence, "Sequence does not contain monomer B, cannot calculate blockiness with reference to B"
+        if 'A' in sequence and 'B' in sequence:
+            AB = sequence.count('AB')
+            BB = sequence.count('BB')
+            BA = sequence.count('BA')
+            AA = sequence.count('AA')
+            if 'AB' in sequence:
+                blockiness = BB/AB
+            else:
+                blockiness = BB/BA
+            #Calculate block length A
+            block_list_A = [x for x in sequence.split('B') if x!='']
+            block_length_A = mean([len(b) for b in block_list_A])
+            #Calculate block length B
+            block_list_B = [x for x in sequence.split('A') if x!='']
+            block_length_B = mean([len(b) for b in block_list_B])
+            return blockiness, block_length_A, block_length_B
+
+        else:  
+            return 'Molecule is not a co-polymer, no blockiness calculation performed', 0, len(sequence)
+    
 
     else:
-        return 'Molecule is not a co-polymer, no blockiness calculation performed', 0, len(sequence)
+        raise ValueError("wrt parameter must be 'A' or 'B'")
 
 def calculate_box_components(chains, monomers, sequence, salt_concentration = 0.0 * unit.mole / unit.liter, residual_monomer = 0.00, solvated=True):
     """
@@ -511,7 +538,7 @@ class polymer_system:
                  stereoisomerism_input=None,
                  terminals='standard', 
                  perc_A_target=100, 
-                 blockiness_target=1.0, 
+                 blockiness_target=[1.0, 'A'], 
[…]
