### 2024-12-11 bb91c5d
debug blockiness spec def
