+++ swiftpol/build.py
@@ -619,7 +619,7 @@ class polymer_system:
         if copolymer==True:
             self.blockiness_target = blockiness_target[0]
             self.A_target = perc_A_target
-            def spec(sequence, blockiness): #Define limits of A percentage and blockiness from input
+            def spec(sequence): #Define limits of A percentage and blockiness from input
                 acceptance_dec = acceptance/100
                 actual_A = (sequence.count('A')/len(sequence))*100
                 blockiness = blockiness_gen(sequence, blockiness_target[1])[0]
@@ -642,8 +642,7 @@ class polymer_system:
             for n in range(num_chains):
                 length_actual = np.random.normal(length_target, 0.5)
                 sequence = reduce(lambda x, y: x + y, np.random.choice(['A', 'B'], size=(int(length_actual),), p=[perc_A_target/100,1-(perc_A_target/100)]))
-                blockiness = blockiness_gen(sequence, blockiness_target[1])[0]
-                if spec(sequence, blockiness)==True:
+                if spec(sequence)==True:
                     if stereoisomerism_input is not None:
                         sequence_stereo = introduce_stereoisomers(stereo_monomer, instance, sequence)
                         pol = build_polymer(sequence=sequence_stereo, monomer_list = monomer_list, reaction = reaction, terminal=terminals, chain_num=n+1)
@@ -664,7 +663,7 @@ class polymer_system:
                     length_actual = np.random.normal(length_target, 0.5)
                     sequence = reduce(lambda x, y: x + y, np.random.choice(['A', 'B'], size=(int(length_actual),), p=[perc_A_target/100,1-(perc_A_target/100)]))
                     blockiness = blockiness_gen(sequence, blockiness_target[1])[0]
-                    if spec(sequence, blockiness)==True:
+                    if spec(sequence)==True:
                         if stereoisomerism_input is not None:
                             sequence_stereo = introduce_stereoisomers(stereo_monomer, instance, sequence) 
                             pol = build_polymer(sequence=sequence_stereo, monomer_list = monomer_list, reaction = reaction, terminal=terminals, chain_num=n+1)
