### 2024-12-11 8d1dafe
debug blockiness spec definition
