### 2024-12-13 6d37035
tests for residue names on ester terminal
