### 2024-12-13 af7bbe9
debug ester terminal build
