### 2024-12-13 49d62be
debug ester terminal ids
