### 2024-12-16 e990fc5
debug blockiness wrt B test
