### 2024-12-21 2fd195b
Removing a redundant assertion in blockiness calc
