### 2024-12-21 46250ba
packmol wrapper for swiftpol ensemble
