### 2024-12-21 d2dea09
docstring and test for self.pack_solvated_system()
