### 2024-12-21 afd397a
missing self on build.build_polymer() class function
