### 2024-12-21 e2777f2
multi solvated chain ensemble option
