### 2024-12-21 8f93e27
debug solavtion class function
