### 2024-12-21 47dccb4
debug solvation class function test
