### 2024-12-27 95bc877
Add polyply file generator to class functions
