### 2024-12-27 b1af769
Residual oligomer calculation
