+++ swiftpol/build.py
@@ -356,8 +356,7 @@ def calculate_box_components(chains, monomers, sequence, salt_concentration = 0.
     from openff.interchange.components._packmol import UNIT_CUBE, pack_box, RHOMBIC_DODECAHEDRON, solvate_topology
     from openff.interchange.components._packmol import _max_dist_between_points, _compute_brick_from_box_vectors, _center_topology_at
     import warnings
-    if not solvated and len(chains) < 15:
-        warnings.warn('Residual monomer calculation may not be accurate for small systems. Please check the output by querying the value of residual_monomer_actual')
+
     #Create molecules for the purpose of mass calculation
     #Water
     water = Molecule.from_smiles('O')
@@ -414,12 +413,11 @@ def calculate_box_components(chains, monomers, sequence, salt_concentration = 0.
     for m in topology.molecules:
         rolling_mass += sum(atom.mass for atom in m.atoms)
     rolling_mass += nacl_mass * nacl_to_add
-    if solvated:
-        rolling_mass += water_mass * water_to_add
-    
-    # residual monomer to add
+
+    # Calculate residual monomer to add
     mass_to_add = (rolling_mass.magnitude/100-residual_monomer) * residual_monomer
-    
+    if mass_to_add < 90:
+        warnings.warn('Residual monomer calculation may not be accurate for small systems, as the residual monomer mass can be lower than the weight of a single monomer. Please check the output by querying the value of residual_monomer_actual')
     
     if 'A' in sequence and 'B' in sequence:
         A_rd = Chem.MolFromSmiles(monomers[0])
@@ -636,6 +634,8 @@ class polymer_system:
         lengths = []
         
         self.monomers = [mono.replace("[I]", "") for mono in monomer_list]
+        self.reaction = reaction
+        reaction = AllChem.ReactionFromSmarts(reaction)
         if stereoisomerism_input is not None:
             stereo_monomer, instance, new_smiles = stereoisomerism_input
             monomer_list.append(new_smiles)
@@ -859,7 +859,7 @@ class polymer_system:
         """
         Generate input files for Polyply from the system.
 
-        This method generates the input files required for [Polyply](https://github.com/marrink-lab/polyply_1.0/) from the system.
+        This method generates the input files required for Polyply (https://github.com/marrink-lab/polyply_1.0/) from the system.
 
         Parameters
         ----------
@@ -919,3 +919,163 @@ class polymer_system:
             return_tuple += (pdb_file,)
 
         return return_tuple
+
+    def calculate_residuals(self, residual_monomer = 0, residual_oligomer = 0):
+        """
+        Generate residual monomer and oligomer molecules, and molecule counts.
+
+        Parameters
+        ----------
+        residual_monomer : float
+            The desired residual monomer concentration in the simulation box. Default is 0.00.
+        residual_oligomer : float
+            The desired residual oligomer concentration in the simulation box. Default is 0.00.
+            
+        Returns
+        -------
+        tuple
+        A tuple containing the following elements:
+            - A list of OpenFF Molecule objects representing the residual monomer and oligomer molecules.
+            - A list of molecule counts, corresponding with the molecule list
+            - Actual value for % residual monomer
+            - Actual value for % residual oligomer
+
+        Notes
+        -----
+        EXPERIMENTAL CAPABILITY. Proceed with caution
+
+        This function works independently of, and is unrelated to, the build.calculate_box_components function.
+        
[…]
