### 2024-12-27 aac1521
remove test for depracated residual monomer calculation
