### 2024-12-27 856cdc2
debug new residuals function
