### 2024-12-28 3952db4
test for new residual oligomer and monomer calculation
