### 2024-12-28 6896d04
test for new residual monomer and oligomer calculation
