### 2025-01-03 8013182
fix bug in residual oligomer code
