### 2025-01-07 991280d
openeye dependency removed
