### 2025-01-07 b2c61df
removed openeye dependency from __init__
