+++ swiftpol/__init__.py
@@ -9,7 +9,13 @@ import random
 import numpy as np
 import numpy as np
 import time
-from openeye import oechem
+try:
+    from openeye import oechem
+    oechem_imported = True
+except:
+    import warnings
+    warnings.warn("OpenEye is not installed. You will not be able to use OpenEye Toolkits for conformer generation.")
+
 
 import openmm
 from openff.toolkit.topology import Molecule, Topology
