### 2025-01-07 884fac5
debug non-OE option for conformer generation
