### 2025-01-07 fa691b6
remove espaloma dependency from __init__
