### 2025-01-15 2a0c12d
fix bug in residual calculation for 0 residual oligo
