+++ swiftpol/_version.py
@@ -1,1 +1,1 @@
-__version__ = "1.0.0+239.g3e7740f.dirty"
+__version__ = "1.0.0+268.g8013182.dirty"
