### 2025-02-06 #7567 @RMeli
@Kevin-Mattheus-Moerman precisely, but unfortunately I have the same COI as in https://github.com/openjournals/joss-reviews/issues/7502. One of the authors is a core developer of MDAnalysis, and so am I.
