### 2025-02-12 db6f3ff
replaced generate residuals with new class function
