### 2025-02-12 5ef1ddf
Fix docstring in calculate_residuals
