### 2025-02-12 df67792
debug docstring and add box_vectors calculation
