+++ Dev_tools/swiftpol.yml
@@ -17,7 +17,7 @@ dependencies:
   - openbabel
   - openeye-toolkits
   - openff-forcefields
-  - openff-interchange>=0.3.27
+  - openff-interchange>=0.4.0
   - openff-nagl>=0.3.8 
   - openff-models
   - openff-toolkit>=0.16.0
@@ -32,7 +32,7 @@ dependencies:
   - pydantic
   - python>=3.11.5
   - python-dateutil
-  - pytorch
+  - pytorch>=2.3.1
   - qcelemental
   - qcengine
   - qcportal
