### 2025-02-12 c030b26
free up more disk space during tests
