+++ .github/workflows/SwiftPol_tests.yml
@@ -9,7 +9,7 @@ on:
 jobs:
   build:
 
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-22.04-large
     strategy:
       fail-fast: false
       matrix:
