+++ .github/workflows/SwiftPol_tests.yml
@@ -9,7 +9,7 @@ on:
 jobs:
   build:
 
-    runs-on: ubuntu-22.04-large
+    runs-on: ubuntu-latest
     strategy:
       fail-fast: false
       matrix:
