### 2025-03-10 a834084
fix mismatch in code defaults and docs
