### 2025-03-26 #7567 @amoeba
Hi @srmnitc, I unfortunately don't have time to review this at the moment.
