### 2025-03-26 #7567 @jfaraudo
Yes, I can do that.
