+++ paper/paper.bib
@@ -382,23 +382,6 @@ Fix misnamed test
 	year = {2016},
 }
 
-@article{hawkins_comparison_2007,
-	title = {Comparison of shape-matching and docking as virtual screening tools},
-	volume = {50},
-	issn = {0022-2623},
-	doi = {10.1021/jm0603365},
-	abstract = {Ligand docking is a widely used approach in virtual screening. In recent years a large number of publications have appeared in which docking tools are compared and evaluated for their effectiveness in virtual screening against a wide variety of protein targets. These studies have shown that the effectiveness of docking in virtual screening is highly variable due to a large number of possible confounding factors. Another class of method that has shown promise in virtual screening is the shape-based, ligand-centric approach. Several direct comparisons of docking with the shape-based tool ROCS have been conducted using data sets from some of these recent docking publications. The results show that a shape-based, ligand-centric approach is more consistent than, and often superior to, the protein-centric approach taken by docking.},
-	language = {eng},
-	number = {1},
-	journal = {Journal of Medicinal Chemistry},
-	author = {Hawkins, Paul C. D. and Skillman, A. Geoffrey and Nicholls, Anthony},
-	month = jan,
-	year = {2007},
-	pmid = {17201411},
-	keywords = {Proteins, Binding Sites, Crystallography, X-Ray, Ligands, Molecular Conformation, Molecular Structure, Protein Binding, Quantitative Structure-Activity Relationship, ROC Curve},
-	pages = {74--82},
-}
-
 @misc{kings_college_london_kings_2024,
 	title = {King's {Computational} {Research}, {Engineering} and {Technology} {Environment} ({CREATE}).},
 	url = {https://doi.org/10.18742/rnvf-m076},
@@ -407,8 +390,8 @@ Fix misnamed test
 	year = {2024},
 }
 
-@article{triandafilidi_molecular_2016,
-	title = {Molecular dynamics simulations of monodisperse/bidisperse polymer melt crystallization},
+@article{triandafilidi_monodmolecular_2016,
+	title = {{monodMolecular} dynamics simulations of monodisperse/bidisperse polymer melt crystallization},
 	volume = {54},
 	copyright = {© 2016 Wiley Periodicals, Inc.},
 	issn = {1099-0488},
@@ -442,4 +425,34 @@ Fix misnamed test
 	year = {2009},
 	pmid = {19229944},
 	pages = {2157--2164},
+}
+
+@misc{openeye_cadence_molecular_sciences_omega_nodate,
+	address = {Santa Fe, NM},
+	title = {{OMEGA} 6.0.0.1.},
+	url = {https://www.eyesopen.com},
+	abstract = {Molecular modeling and cheminformatics software tools built in the cloud to accelerate lead discovery and optimization in drug development.},
+	urldate = {2025-04-02},
+	publisher = {OpenEye, Cadence Molecular Sciences},
+	author = {{OpenEye, Cadence Molecular Sciences}},
+	file = {Snapshot:/Users/hannahturney/Zotero/storage/HHSUE95L/www.eyesopen.com.html:text/html},
+}
+
+@article{hawkins_conformer_2010,
+	title = {Conformer {Generation} with {OMEGA}: {Algorithm} and {Validation} {Using} {High} {Quality} {Structures} from the {Protein} {Databank} and {Cambridge} {Structural} {Database}},
+	volume = {50},
+	issn = {1549-9596},
+	shorttitle = {Conformer {Generation} with {OMEGA}},
+	url = {https://doi.org/10.1021/ci100031x},
+	doi = {10.1021/ci100031x},
[…]
