+++ paper/paper.bib
@@ -427,7 +427,7 @@ Fix misnamed test
 	pages = {2157--2164},
 }
 
-@misc{openeye_cadence_molecular_sciences_omega_nodate,
+@misc{openeye_cadence_molecular_sciences_omega_2025,
 	address = {Santa Fe, NM},
 	title = {{OMEGA} 6.0.0.1.},
 	url = {https://www.eyesopen.com},
@@ -435,6 +435,8 @@ Fix misnamed test
 	urldate = {2025-04-02},
 	publisher = {OpenEye, Cadence Molecular Sciences},
 	author = {{OpenEye, Cadence Molecular Sciences}},
+	month = feb,
+	year = {2025},
 	file = {Snapshot:/Users/hannahturney/Zotero/storage/HHSUE95L/www.eyesopen.com.html:text/html},
 }
 
