+++ paper/paper.bib
@@ -434,7 +434,7 @@ Fix misnamed test
 	abstract = {Molecular modeling and cheminformatics software tools built in the cloud to accelerate lead discovery and optimization in drug development.},
 	urldate = {2025-04-02},
 	publisher = {OpenEye, Cadence Molecular Sciences},
-	author = {{OpenEye, Cadence Molecular Sciences}},
+	author = {{OpenEye Cadence Molecular Sciences}},
 	month = feb,
 	year = {2025},
 	file = {Snapshot:/Users/hannahturney/Zotero/storage/HHSUE95L/www.eyesopen.com.html:text/html},
