### 2025-04-09 f941251
Updates to PDI calculation
