+++ swiftpol/build.py
@@ -9,6 +9,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 import time
+from collections import Counter
 try:
     from openeye import oechem
     oechem_imported = True
