### 2025-04-09 e62cafe
debug build_tests missing conformer
