+++ swiftpol/build.py
@@ -1082,3 +1082,247 @@ class polymer_system:
             number_of_copies = [A_to_add, B_to_add] + [1 for c in range(len(oligomers))]
     
         return molecules, number_of_copies, residual_monomer_actual, residual_oligomer_actual
+
+class polymer_system_from_PDI:
+    """
+    Build a polymer system based on a target polydispersity index (PDI).
+
+    This class generates a polymer system with a specified number of chains, 
+    target chain length, and target PDI. It supports the generation of copolymers, 
+    stereoisomers, and blockiness control. The system is built using OpenFF and RDKit 
+    toolkits, with optional support for OpenEye toolkits if installed.
+
+    **Note**: This is an experimental capability and is not routinely tested. 
+    Proceed with caution.
+
+    Parameters
+    ----------
+    monomer_list : list of str
+        A list of monomer SMILES strings to be used in the polymer system.
+    reaction : str
+        A SMARTS string defining the polymerization reaction.
+    length_target : int
+        The target average chain length for the polymer system.
+    num_chains : int
+        The number of polymer chains to generate.
+    PDI_target : float
+        The target polydispersity index (PDI) for the polymer system.
+    stereoisomerism_input : tuple, optional
+        A tuple containing the stereoisomer monomer, the fraction of stereoisomers 
+        to introduce, and the SMILES string of the stereoisomer. Default is None.
+    terminals : str, optional
+        The type of terminal groups to use for the polymer chains. Default is 'standard'.
+    perc_A_target : float, optional
+        The target percentage of monomer 'A' in the copolymer. Default is 100.
+    blockiness_target : list, optional
+        A list containing the target blockiness value and the monomer to control 
+        blockiness for. Default is [1.0, 'A'].
+    copolymer : bool, optional
+        Whether to generate a copolymer. Default is False.
+    acceptance : float, optional
+        The acceptance margin (in percentage) for the target blockiness and 
+        monomer percentage. Default is 10.
+
+    Attributes
+    ----------
+    chains : list of openff.toolkit.topology.Molecule
+        A list of OpenFF Molecule objects representing the polymer chains.
+    chain_rdkit : list of rdkit.Chem.Mol
+        A list of RDKit Molecule objects representing the polymer chains.
+    sequence : str
+        The sequence of monomers in the polymer system.
+    mol_weight_average : float
+        The average molecular weight of the polymer chains.
+    PDI : float
+        The polydispersity index of the polymer system.
+    Mn : float
+        The number-average molecular weight of the polymer system.
+    Mw : float
+        The weight-average molecular weight of the polymer system.
+    num_chains : int
+        The number of polymer chains in the system.
+    length_average : float
+        The average chain length of the polymer system.
+    lengths : list of int
+        A list of chain lengths for the polymer chains.
+    min_length : int
+        The minimum chain length in the polymer system.
+    max_length : int
+        The maximum chain length in the polymer system.
+    A_actual : float or None
+        The actual percentage of monomer 'A' in the copolymer. None if not a copolymer.
+    perc_A_actual : list of float or None
+        A list of the actual percentages of monomer 'A' in each chain. None if not a copolymer.
+    B_block_length : float or None
+        The average block length of monomer 'B'. None if not a copolymer.
+    A_block_length : float or None
+        The average block length of monomer 'A'. None if not a copolymer.
+    blockiness_list : list of float or None
[…]
