### 2025-04-09 403cb8e
fixed PDI for build tests
