### 2025-04-09 364d88b
debug ensemble from PDI function
