### 2025-04-10 04dad74
New build tests for system from PDI function
