+++ tests/build_tests.py
@@ -366,7 +366,7 @@ class TestPolymerSystem(unittest.TestCase):
 class TestPolymerSystemFromPDI(unittest.TestCase):
     def test_init(self):
         #Test PDI=1.7
-        sys = polymer_system_from_PDI(monomer_list=['OC(=O)COI'], 
+        sys = build.polymer_system_from_PDI(monomer_list=['OC(=O)COI'], 
                             reaction = '[C:1][O:2][H:3].[I:4][O:5][C:6]>>[C:1][O:2][C:6].[H:3][O:5][I:4]', 
                             length_target= 50, 
                             num_chains=50,
@@ -377,7 +377,7 @@ class TestPolymerSystemFromPDI(unittest.TestCase):
         self.assertTrue(45 <= round(sys.max_length)<= 55)
         self.assertTrue(1.2 <= sys.PDI <= 2.3)
         #Test PDI=1.0 (monodisperse)
-        sys = polymer_system_from_PDI(monomer_list=['OC(=O)COI'], 
+        sys = build.polymer_system_from_PDI(monomer_list=['OC(=O)COI'], 
                             reaction = '[C:1][O:2][H:3].[I:4][O:5][C:6]>>[C:1][O:2][C:6].[H:3][O:5][I:4]', 
                             length_target= 50, 
                             num_chains=50,
