### 2025-04-10 f20a40b
test for copolymer ensemble from PDI
