+++ README.md
@@ -51,112 +51,9 @@ We acknowledge the use of a free academic license provided by OpenEye, Candence 
 ├── CODE_OF_CONDUCT.md
 ├── Dev_tools
 │   └── swiftpol.yml
-├── docs
-│   ├── api.rst
-│   ├── _build
-│   │   ├── doctrees
-│   │   │   ├── api.doctree
-│   │   │   ├── environment.pickle
-│   │   │   ├── getting_started.doctree
-│   │   │   └── index.doctree
-│   │   └── html
-│   │       ├── api.html
-│   │       ├── genindex.html
-│   │       ├── getting_started.html
-│   │       ├── index.html
-│   │       ├── _modules
-│   │       │   ├── index.html
-│   │       │   └── swiftpol
-│   │       │       ├── build.html
-│   │       │       └── parameterize.html
-│   │       ├── objects.inv
-│   │       ├── py-modindex.html
-│   │       ├── search.html
-│   │       ├── searchindex.js
-│   │       ├── _sources
-│   │       │   ├── api.rst.txt
-│   │       │   ├── getting_started.rst.txt
-│   │       │   └── index.rst.txt
-│   │       └── _static
-│   │           ├── basic.css
-│   │           ├── css
-│   │           │   ├── badge_only.css
-│   │           │   ├── fonts
-│   │           │   │   ├── fontawesome-webfont.eot
-│   │           │   │   ├── fontawesome-webfont.svg
-│   │           │   │   ├── fontawesome-webfont.ttf
-│   │           │   │   ├── fontawesome-webfont.woff
-│   │           │   │   ├── fontawesome-webfont.woff2
-│   │           │   │   ├── lato-bold-italic.woff
-│   │           │   │   ├── lato-bold-italic.woff2
-│   │           │   │   ├── lato-bold.woff
-│   │           │   │   ├── lato-bold.woff2
-│   │           │   │   ├── lato-normal-italic.woff
-│   │           │   │   ├── lato-normal-italic.woff2
-│   │           │   │   ├── lato-normal.woff
-│   │           │   │   ├── lato-normal.woff2
-│   │           │   │   ├── Roboto-Slab-Bold.woff
-│   │           │   │   ├── Roboto-Slab-Bold.woff2
-│   │           │   │   ├── Roboto-Slab-Regular.woff
-│   │           │   │   └── Roboto-Slab-Regular.woff2
-│   │           │   └── theme.css
-│   │           ├── doctools.js
-│   │           ├── documentation_options.js
-│   │           ├── file.png
-│   │           ├── fonts
-│   │           │   ├── Lato
-│   │           │   │   ├── lato-bold.eot
-│   │           │   │   ├── lato-bolditalic.eot
-│   │           │   │   ├── lato-bolditalic.ttf
-│   │           │   │   ├── lato-bolditalic.woff
-│   │           │   │   ├── lato-bolditalic.woff2
-│   │           │   │   ├── lato-bold.ttf
-│   │           │   │   ├── lato-bold.woff
-│   │           │   │   ├── lato-bold.woff2
-│   │           │   │   ├── lato-italic.eot
-│   │           │   │   ├── lato-italic.ttf
-│   │           │   │   ├── lato-italic.woff
-│   │           │   │   ├── lato-italic.woff2
-│   │           │   │   ├── lato-regular.eot
-│   │           │   │   ├── lato-regular.ttf
-│   │           │   │   ├── lato-regular.woff
-│   │           │   │   └── lato-regular.woff2
-│   │           │   └── RobotoSlab
[…]
