### 2025-04-14 fc2f762
Added exception for bad reactions
