### 2025-04-15 67c1bd2
Added three PEGDA building functions.
