+++ swiftpol/demo.py
@@ -528,3 +528,181 @@ class PLGA_system:
                                 center_solute='BRICK')
         
         return bulk_system
+    
+def build_starting_polymer(sequence='ABBBBBBBBBA', 
+                           monomers=["IC=CC(=O)OI", "IOCCOI"], 
+                           polymer_reaction=AllChem.ReactionFromSmarts("[C:1]-[O:2]-[I:3].[C:4]-[O:5]-[I:6]>>[C:1]-[O:2]-[C:4].[I:3]-[I:6].[O:5]"), 
+                           activation_reaction=AllChem.ReactionFromSmarts('[C:1]=[C:2].[Cl:3].[I:4]>>[C:1](-[I:4])-[C:2]-[Cl:3]')):
+    """
+    Builds a polymer and adds reactive sites.
+
+    Parameters:
+        - sequence (str): Polymer sequence
+        - monomers (list): List of monomers (SMILES format)
+        - polymer_reaction (rdkit reaction): Reaction to form the polymer
+        - activation_reaction (rdkit reaction): Reaction to introduce reactive sites
+
+    Returns:
+        - starting_polymer (rdkit Mol): PEGDA polymer with reactive sites
+    """
+    from swiftpol import build
+    #Build PEGDA polymer
+    polymer = build.build_polymer(sequence=sequence, monomer_list=monomers, reaction=polymer_reaction)
+
+    activatorCl = Chem.MolFromSmiles('Cl')
+    activatorI = Chem.MolFromSmiles('I')
+
+    # Apply activation reaction twice
+    starting_polymer = activation_reaction.RunReactants([polymer, activatorCl, activatorI])[0][0]
+    Chem.SanitizeMol(starting_polymer)
+
+    starting_polymer = activation_reaction.RunReactants([starting_polymer, activatorCl, activatorI])[0][0]
+    Chem.SanitizeMol(starting_polymer)
+
+    Chem.AddHs(starting_polymer)
+    return starting_polymer
+
+def replace_halogens_with_hydrogens(mol):
+    """
+    Replaces all halogen atoms (F, Cl, Br, I) in an RDKit molecule with hydrogen atoms.
+
+    This function iterates through the atoms in the given molecule, identifies halogen atoms 
+    (fluorine, chlorine, bromine, iodine), removes them, and replaces them with explicit hydrogens. 
+    The resulting molecule is sanitized and updated to ensure chemical validity.
+
+    Parameters
+    ----------
+    mol : rdkit.Chem.Mol
+        The input RDKit molecule from which halogens will be replaced with hydrogens.
+
+    Returns
+    -------
+    rdkit.Chem.Mol
+        A new RDKit molecule with all halogens replaced by explicit hydrogens.
+
+    Notes
+    -----
+    - The function uses an editable RDKit molecule (RWMol) to modify the structure.
+    - The molecule is sanitized after halogen removal to ensure chemical correctness.
+    - Explicit hydrogens are added to the resulting molecule.
+    """
+    editable = Chem.RWMol(mol)   #Creates an editable version of an RDKit molecule
+    halogen_atomic_nums = [9, 17, 35, 53]  #F, Cl, Br, I
+
+    for atom in list(editable.GetAtoms()):  #list() to avoid issues
+        if atom.GetAtomicNum() in halogen_atomic_nums:
+            idx = atom.GetIdx()
+            #Remove the halogen
+            editable.RemoveAtom(idx)
+
+    #Sanitize and add hydrogens back
+    mol_no_halogens = editable.GetMol()
+    mol_no_halogens.UpdatePropertyCache()
+    Chem.SanitizeMol(mol_no_halogens)
+
+    #Add explicit hydrogens
+    mol_with_h = Chem.AddHs(mol_no_halogens)
+    return mol_with_h
+
[…]
