### 2025-04-23 matta-research-group/SwiftPol#14 @ricalessandri
[JOSS review] 1. Text reptition
This text appears twice in the JOSS article current proof:

"Here, we will detail the development of SwiftPol - a user-guided Python tool for building repre-
45 sentative polymer ensembles, and subsequent studies to show its relevance and performance."

This is a comment related to my review here https://github.com/openjournals/joss-reviews/issues/8053
