### 2025-04-23 matta-research-group/SwiftPol#15 @ricalessandri
[JOSS review] 2. Packmol citation
Packmol should be cited at:

"... ensemble to existing tools such as packmol [CITE] to generate ..."

Moreover, it's not clear while the packmol paper is cited at this point:

"... using built-in swiftpol functions (Martinez et al)."

This should be clarified.

This issue concerns my review at https://github.com/openjournals/joss-reviews/issues/8053
