### 2025-04-24 8ee7b8b
Update paper.md to address reviewer comments
Addressing comments  in issue #15 and #14 - corrected Packmol citation and text repetition
