### 2025-04-24 #8053 @hannaomi
Addressed issues raised by @ricalessandri in most recent commits to [paper branch](https://github.com/matta-research-group/SwiftPol/tree/paper).[ New proof available](https://github.com/matta-research-group/SwiftPol/actions/runs/14646636584/artifacts/3003599308).
