### 2025-04-24 matta-research-group/SwiftPol#15 @hannaomi
Change committed to [paper branch](https://github.com/matta-research-group/SwiftPol/tree/paper). Packmol citation corrected at "... ensemble to existing tools such as packmol (Martinez et al) to generate ..."

and removed from latter sentence.
