### 2025-04-24 matta-research-group/SwiftPol#14 @hannaomi
Change committed to [paper branch](https://github.com/matta-research-group/SwiftPol/tree/paper).

Text repetition removed
