### 2025-04-25 ce00851
Improve installation instructions in README.md
