+++ Dev_tools/swiftpol.yml
@@ -20,7 +20,7 @@ dependencies:
   - openff-interchange>=0.4.0
   - openff-nagl>=0.3.8 
   - openff-models
-  - openff-toolkit>=0.16.0
+  - openff-toolkit>=0.16.6
   - openff-toolkit-examples
   - openff-units
   - openff-utilities
@@ -40,7 +40,7 @@ dependencies:
   - seaborn
   - smirnoff99frosst
   - statsmodels
-  - dgl>=2.0.0
+  - dgl==2.0.0
   - pip
   - pip:
       - amberutils
@@ -48,6 +48,6 @@ dependencies:
       - pdb4amber
       - pytraj
       - rdkit
-      - torch >= 2.0.0
+      #- torch == 1.0.0
       - torchdata
 
