### 2025-04-25 2e99e37
Updated installation instructions in README.md
