+++ Dev_tools/swiftpol.yml
@@ -32,7 +32,7 @@ dependencies:
   - pydantic
   - python>=3.11.5
   - python-dateutil
-  - pytorch==2.0.0
+  - pytorch>=2.0.0
   - qcelemental
   - qcengine
   - qcportal
@@ -48,6 +48,6 @@ dependencies:
       - pdb4amber
       - pytraj
       - rdkit
-      - torch == 2.0.0
+      #- torch == 2.0.0
       - torchdata
 
