### 2025-04-25 59ef1fd
fix bug in build_tests
