### 2025-04-25 8358b56
fix small bug in build_tests
