### 2025-04-25 1265156
fix small bug in build_tests
