### 2025-04-25 d4b0caa
Fix error in docs compiling and futureproof install
