+++ docs/api.rst
@@ -16,12 +16,20 @@ swiftpol.build
    :undoc-members:
    :show-inheritance:
    :no-imported-members:
-   :exclude-members: polymer_system
+   :exclude-members: polymer_system, polymer_system_from_PDI
 
 swiftpol.build.polymer_system
 -----------------------------
 
 .. autoclass:: swiftpol.build.polymer_system
+   :members: 
+   :special-members: __init__
+   :exclude-members: build_linear_copolymer, build_polymer, introduce_stereoisomers, calculate_box_components, PDI, blockiness_gen, Interchange, ExactMolWt, OpenEyeToolkitWrapper, RDKitToolkitWrapper, mean, np, oechem, reduce, unit, AllChem, UNIT_CUBE, pack_box
+
+swiftpol.build.polymer_system_from_PDI
+-----------------------------
+
+.. autoclass:: swiftpol.build.polymer_system_from_PDI
    :members: 
    :special-members: __init__
    :exclude-members: build_linear_copolymer, build_polymer, introduce_stereoisomers, calculate_box_components, PDI, blockiness_gen, Interchange, ExactMolWt, OpenEyeToolkitWrapper, RDKitToolkitWrapper, mean, np, oechem, reduce, unit, AllChem, UNIT_CUBE, pack_box
