### 2025-04-25 ce7b072
Improve formatting on README.md installation
