### 2025-04-25 #8053 @hannaomi
@jfaraudo Please install from the main branch. I have just checked that all example notebooks run with both main branch and paper branch install. If you are still running into problems with the examples using the main branch install,  please try a fresh clone of the repo.

Many thanks,
Hannah
