### 2025-04-25 dcb8892
Updated README to include detailed depenedencies
