### 2025-04-25 8c60e5c
Improved dependency requirements in readme and paper
