### 2025-04-25 0a3cf9c
fix error compiling dependency bullets
