+++ paper/paper.md
@@ -162,14 +162,22 @@ n_{chains} = \mbox{total number of chains built}
 [YAML environment file listing all required packages to use SwiftPol](https://github.com/matta-research-group/SwiftPol/blob/main/Dev_tools/swiftpol.yml)
 
 To use SwiftPol please download the following packages:
+
 - RDkit
+
 - openff-interchange
+
 - openff-toolkit
+
 - openff-nagl
+
 - openeye-toolkits
+
 - dgl==2.0.0
 
+
 optional dependencies:
+
 - espaloma charge
 
 
