### 2025-04-28 1acde58
fixed error in percentage A calculation for polymer system
