+++ Dev_tools/swiftpol.yml
@@ -17,12 +17,12 @@ dependencies:
   - openbabel
   - openeye-toolkits
   - openff-forcefields
-  - openff-interchange>=0.4.0
+  - openff-interchange
   - openff-nagl>=0.3.8 
   - openff-models
-  - openff-toolkit>=0.16.0
+  - openff-toolkit==0.16.6
   - openff-toolkit-examples
-  - openff-units
+  - openff-units==0.2.2
   - openff-utilities
   - openmm
   - packmol
