### 2025-04-28 4382105
fix upenff-units dependency in quick install instructions
