### 2025-04-28 d3666d5
fix openff-units dependency issue in paper and quick install
