### 2025-04-28 9a7ff1e
Updated dockerfile to include dedicated docker.yml file for swiftpol
