### 2025-05-01 6841d34
residual oligomer actual self-assignment
