### 2025-05-05 6a195a5
Add class functions to polymer_system_from_PDI
