+++ swiftpol/build.py
@@ -1262,11 +1262,12 @@ class polymer_system_from_PDI:
         if copolymer==True:
             n=0
             for l in chain_lengths:
-                sequence = reduce(lambda x, y: x + y, np.random.choice(['A', 'B'], size=(l,), p=[perc_A_target/100,1-(perc_A_target/100)]))
-                blockiness = blockiness_gen(sequence, blockiness_target[1])[0]
-                while spec(sequence)==False:
+                for i in range(1000):
                     sequence = reduce(lambda x, y: x + y, np.random.choice(['A', 'B'], size=(l,), p=[perc_A_target/100,1-(perc_A_target/100)]))
                     blockiness = blockiness_gen(sequence, blockiness_target[1])[0]
+                    if spec(sequence)==False:
+                        sequence = reduce(lambda x, y: x + y, np.random.choice(['A', 'B'], size=(l,), p=[perc_A_target/100,1-(perc_A_target/100)]))
+                        blockiness = blockiness_gen(sequence, blockiness_target[1])[0]
 
                 if stereoisomerism_input is not None:
                     sequence_stereo = introduce_stereoisomers(stereo_monomer, instance, sequence)
@@ -1282,9 +1283,6 @@ class polymer_system_from_PDI:
                 BBL.append(blockiness_gen(sequence, blockiness_target[1])[1])
                 ABL.append(blockiness_gen(sequence, blockiness_target[1])[2])
                 n+=1
-
-
-
             self.B_block_length = mean(BBL)
             self.A_block_length = mean(ABL)
             self.blockiness_list = blockiness_list
@@ -1329,8 +1327,360 @@ class polymer_system_from_PDI:
         self.max_length = max(lengths)
         print('System built!, size =', self.num_chains)
         
-    def __repr__(self):
+        def __repr__(self):
                 
         description = (f"SwiftPol ensemble of size {self.num_chains}, "
                        f"average chain length = {self.length_average}-mers, PDI = {self.PDI}")
         return description
+
+    def generate_conformers(self):
+        """
+        Generate conformers for each polymer chain in the system.
+
+        This method uses the OpenFF toolkit OpenEye Wrapper to generate conformers for each polymer chain in the system.
+        It first checks if the OpenEye toolkit is licensed and available. If it is, it uses the OpenEyeToolkitWrapper
+        to generate conformers. Otherwise, it falls back to using the RDKitToolkitWrapper. Each chain is processed
+        to generate a single conformer, and unique atom names are assigned to each chain.
+
+        Raises
+        ------
+        ImportError
+            If neither RDKit nor OpenEye toolkits are available.
+        """
+
+        from openff.toolkit.utils.toolkits import RDKitToolkitWrapper, OpenEyeToolkitWrapper
+        # Generate conformers using OpenFF toolkit wrapper
+        for chain in self.chains:
+            num = self.chains.index(chain)
+            object = RDKitToolkitWrapper()
+            if oechem_imported:
+                if oechem.OEChemIsLicensed():
+                    object = OpenEyeToolkitWrapper()
+            object.generate_conformers(molecule=chain, n_conformers=1)
+            chain.generate_unique_atom_names()
+            self.chains[num] = chain
+    
+    def charge_system(self, charge_scheme):
+        """
+        Assign partial charges to each polymer chain in the system.
+
+        This method uses one of AM1-BCC, Espaloma, or OpenFF NAGL to assign partial charges to each polymer chain in the system.
+        It iterates over each chain in the `self.chains` list and assigns partial charges to the chain.
+
+        Parameters
+        ----------
+        charge_scheme : str
+            The charge assignment scheme to use. Options are 'AM1_BCC', 'espaloma', or 'NAGL'.
+
+        Raises
+        ------
+        ImportError
[…]
