### 2025-05-05 d6bb20d
Debug docs compiling
