### 2025-05-05 c9daa77
debug build polymer system from PDI error with mean blockiness calculation
