### 2025-05-05 matta-research-group/SwiftPol#20 @ricalessandri
[JOSS review] 3. Documentation needs a statement of need
According to the reviewer checklist (https://github.com/openjournals/joss-reviews/issues/8053#issuecomment-2821201908).
and indeed I think the first page of the documentation (https://matta-research-group.github.io/SwiftPol/index.html) would benefit from 2-3 sentences stating what swiftpol is about
