### 2025-05-05 matta-research-group/SwiftPol#21 @ricalessandri
[JOSS review] 4. Add missing community guidelines
"Community guidelines: Are there clear guidelines for third parties wishing to:
- [x] 1. Contribute to the software **OK**
- [x] 2. Report issues or problems with the software **MISSING**
- [x] 3. Seek support **MISSING**

This concerns my review at: https://github.com/openjournals/joss-reviews/issues/8053#issuecomment-2821201908
