### 2025-05-05 matta-research-group/SwiftPol#22 @ricalessandri
[JOSS review] 5. Other software
This sentence:

"Open-source software packages designed to build in silico polymer chains are focused on the design of polymers at the monomer and single-chain scale"

is not accurate and should be revised.
There's a number of software open source packages that can build *multi-chain* polymer systems:
- polyply [repo](https://github.com/marrink-lab/polyply_1.0); [paper](https://www.nature.com/articles/s41467-021-27627-4)
- radonpy [repo](https://github.com/RadonPy/RadonPy); [paper](https://www.nature.com/articles/s41524-022-00906-4)
- hoobas [repo](https://bitbucket.org/NUaztec/hoobas); [paper](https://www.sciencedirect.com/science/article/abs/pii/S0927025619302745) (by the way, i'm not very familiar with this one but it also seems to stress *polydispersity*)
- there might be others...
