### 2025-05-05 #8053 @ricalessandri
Only these 3 comments left from my side https://github.com/matta-research-group/SwiftPol/issues/20, https://github.com/matta-research-group/SwiftPol/issues/21, https://github.com/matta-research-group/SwiftPol/issues/22.
