### 2025-05-06 ea196b4
remove reaction hard coding in crosslinked demo
