+++ _sources/getting_started.rst.txt
@@ -1,0 +1,61 @@
+Getting Started with Swiftpol
+=============================
+
+Welcome to the Swiftpol documentation! This guide will help you get started with using Swiftpol for polymer molecular dynamics simulations.
+
+
+About SwiftPol
+--------------
+
+A polymer sample contains a natural degree of variation in its structure and non-uniformity between its chains, which is often disregarded in MD systems.
+
+SwiftPol uses a statistical approach to build polydisperse polymer systems, allowing for a more realistic representation of polymer systems in MD.
+
+SwiftPol.build.polymer_system takes as an input: 
+- The simplified molecular-input line-entry system (SMILES) string of all co-monomers.
+- Values of the target average properties of the ensemble: monomer % composition (for copolymers), length, number of chains, blockiness (for blocky copolymers), terminals, residual monomer. 
+- Reaction SMARTS which describes the polymerization reaction associated with their polymer chemistry.
+
+...and build a polydisperse chain ensemble.
+
+
+Installation
+------------
+
+To install Swiftpol, you can use `conda`. Run the following commands in your terminal:
+
+.. code-block:: bash
+
+    git clone https://github.com/matta-research-group/SwiftPol
+    cd swiftpol
+    conda env create -f Dev_tools/swiftpol.yml -n swiftpol -y 
+    conda activate swiftpol
+
+Swiftpol is in development. To install an editable version, clone the repository and install it locally:
+
+.. code-block:: bash
+
+    git clone https://github.com/matta-research-group/SwiftPol
+    cd swiftpol
+    pip install -e .
+
+Basic Usage
+-----------
+
+Example usage can be found in our `example notebooks <https://github.com/matta-research-group/SwiftPol/tree/main/Example_Notebooks>`_.
+
+Contributing
+------------
+
+We welcome contributions from the community. All contributors to SwiftPol should follow our `Code of Conduct <https://github.com/matta-research-group/SwiftPol/tree/main/CODE_OF_CONDUCT.md>`_.
+To submit your feature to be incorporated to the main branch, you should submit a Pull Request. The repository maintainers will review your pull request before accepting your changes.
+
+Support
+-------
+
+If you encounter any issues or have any questions, please open an issue on our `GitHub repository <https://github.com/matta-research-group/SwiftPol/issues>`_.
+
+License
+-------
+
+Swiftpol is licensed under the `BSD-3-Clause License <https://github.com/matta-research-group/SwiftPol/tree/main/LICENSE>`_.
