+++ _sources/api.rst.txt
@@ -1,0 +1,45 @@
+swiftpol
+========
+
+.. automodule:: swiftpol
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :no-imported-members:
+
+
+swiftpol.build
+==============
+
+.. automodule:: swiftpol.build
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :no-imported-members:
+   :exclude-members: polymer_system, polymer_system_from_PDI
+
+swiftpol.build.polymer_system
+-----------------------------
+
+.. autoclass:: swiftpol.build.polymer_system
+   :members: 
+   :special-members: __init__
+   :exclude-members: build_linear_copolymer, build_polymer, introduce_stereoisomers, calculate_box_components, PDI, blockiness_gen, Interchange, ExactMolWt, OpenEyeToolkitWrapper, RDKitToolkitWrapper, mean, np, oechem, reduce, unit, AllChem, UNIT_CUBE, pack_box
+
+swiftpol.build.polymer_system_from_PDI
+-----------------------------
+
+.. autoclass:: swiftpol.build.polymer_system_from_PDI
+   :members: 
+   :special-members: __init__
+   :exclude-members: build_linear_copolymer, build_polymer, introduce_stereoisomers, calculate_box_components, PDI, blockiness_gen, Interchange, ExactMolWt, OpenEyeToolkitWrapper, RDKitToolkitWrapper, mean, np, oechem, reduce, unit, AllChem, UNIT_CUBE, pack_box
+
+
+swiftpol.parameterize
+=====================
+
+.. automodule:: swiftpol.parameterize
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :no-imported-members:
