+++ _sources/getting_started.rst.txt
@@ -12,8 +12,11 @@ A polymer sample contains a natural degree of variation in its structure and non
 SwiftPol uses a statistical approach to build polydisperse polymer systems, allowing for a more realistic representation of polymer systems in MD.
 
 SwiftPol.build.polymer_system takes as an input: 
+
 - The simplified molecular-input line-entry system (SMILES) string of all co-monomers.
+
 - Values of the target average properties of the ensemble: monomer % composition (for copolymers), length, number of chains, blockiness (for blocky copolymers), terminals, residual monomer. 
+
 - Reaction SMARTS which describes the polymerization reaction associated with their polymer chemistry.
 
 ...and build a polydisperse chain ensemble.
