### 2025-05-06 ad94872
Added support and issues reporting advice to README
