### 2025-05-06 matta-research-group/SwiftPol#21 @hannaomi
Hi Riccardo,

Thanks for raising this. Addressed the missing guidelines in commit [ad94872](https://github.com/matta-research-group/SwiftPol/commit/ad9487269d8328d58a11b1ef845336f6495bcc19).
