### 2025-05-06 matta-research-group/SwiftPol#20 @hannaomi
Hi Riccardo,

Addressed this in commit [f1a40ad](https://github.com/matta-research-group/SwiftPol/commit/f1a40adfe5695664db51c2a36bbd65e4395fe873).

Please see new docs at https://matta-research-group.github.io/SwiftPol/index.html
