### 2025-05-06 0b3cd4f
update docs installation instructions
