+++ docs/index.rst
@@ -28,11 +28,11 @@ SwiftPol.build.polymer_system takes as an input:
 ...and builds a polydisperse chain ensemble.
 
 .. image:: _static/melt.png
-   :alt: SwiftPol Logo
+   :alt: Polymer melt
    :width: 300px
    :align: left
 
-above - A 3D representation of a polydisperse polymer system built with SwiftPol. [PolyPly](https://github.com/marrink-lab/polyply_1.0/tree/master/polyply/src) used for generating 3D configuration.
+above - A 3D representation of a polydisperse polymer system built with SwiftPol. `PolyPly <https://github.com/marrink-lab/polyply_1.0/tree/master/polyply/src>`_ used for generating 3D configuration.
 
 
 .. toctree::
