### 2025-05-08 7dd4ed3
Update to paper to address reviewer comments at issue #22
