+++ paper/paper.md
@@ -31,7 +31,10 @@ A polymer sample contains a natural degree of variation in its structure and non
 # Statement of need 
 MD simulations of polymers are often performed with uniform idealized systems that do not capture the heterogeneity of their experimental counterparts. The result of this misalignment is non-convergence between MD-derived polymer properties and experimental data, and these MD simulations can overlook key components of polymer physics such as polydispersity and semi-crystallinity  [@schmid_understanding_2023; @triandafilidi_monodmolecular_2016]. Studies have demonstrated that bulk polymer material properties such as glass transition temperature, hydrophobicity, and inherent viscosity are highly sensitive to variations in polydispersity, making it essential to account for this heterogeneity to capture the true physics of polymer systems [@li_influence_2016; @wan_effect_2021; @ochi_influence_2021]. 
 Polymer MD studies showcase an assortment of approaches to manually incorporate polydispersity into their polymer chain builds [@andrews_structure_2020; @kawagoe_construction_2019; @stipa_molecular_2021]. Although effective for their associated applications, these manual approaches are not universally applicable to different polymer chemistries or are performed using proprietary software.
-Open-source software packages designed to build *in silico* polymer chains are focused on the design of polymers at the monomer and single-chain scale [@davel_parameterization_2024; @klein_hierarchical_2016; @santana-bonilla_modular_2023]. However, there is not currently a software package available that integrates these smaller-scale characteristics into computational polymer models, whilst effectively capturing the heterogeneity and polydispersity of real-life samples. The development of SwiftPol was driven by the need to fill this gap in multi-scale building functionality of existing polymer building packages, to enable the simulation of realistic polymer models.
+
+Open-source software packages designed to build *in silico* polymer chains are focused on the design of polymers at the monomer and single-chain scale [@davel_parameterization_2024; @klein_hierarchical_2016; @santana-bonilla_modular_2023]. 
+
+Packages that have the capability to build multi-chain systems such as Polyply [@grunewald_polyply_2022], RadonPy [@hayashi_radonpy_2022], Polymer Structure Predictor [sahu_polymer_2022], do not contain in-built functions to incorporate molecular weight diversity, and put the onus on the user to generate an ensemble polydisperse chain sequences prior to using their chain building functions. Polymatic [@abbott_polymatic_2013], an algorithm to simulate polymerization, can theoretically create polydisperse systems through its random bond formation scheme, but does not have to ability to simultaneously control other key attributes of polymers such as blockiness and monomer ratio. The python tool Hoobas [@girard_hoobas_2019] builds polydisperse systems of randomized polymer chains for coarse-grained models. However, there is not currently a software package available that contains in-built functions to integrate multiple smaller-scale characteristics (monomer ratio, blockiness, degree of polymerization, chain terminal) into \textbf{atomistic} computational polymer models, whilst effectively capturing the heterogeneity and polydispersity of real-life samples. The development of SwiftPol was driven by the need to fill this gap in multi-scale building functionality of existing polymer building packages, to enable the simulation of realistic polymer models whilst allowing close user control of multiple system characteristics.
 
 
[…]
