+++ paper/paper.bib
@@ -458,3 +458,103 @@ Fix misnamed test
 	pages = {572--584},
 	file = {Full Text PDF:/Users/hannahturney/Zotero/storage/8XR2VFMI/Hawkins et al. - 2010 - Conformer Generation with OMEGA Algorithm and Validation Using High Quality Structures from the Pro.pdf:application/pdf},
 }
+
+@article{grunewald_polyply_2022,
+	title = {Polyply; a python suite for facilitating simulations of macromolecules and nanomaterials},
+	volume = {13},
+	copyright = {2022 The Author(s)},
+	issn = {2041-1723},
+	url = {https://www.nature.com/articles/s41467-021-27627-4},
+	doi = {10.1038/s41467-021-27627-4},
+	abstract = {Molecular dynamics simulations play an increasingly important role in the rational design of (nano)-materials and in the study of biomacromolecules. However, generating input files and realistic starting coordinates for these simulations is a major bottleneck, especially for high throughput protocols and for complex multi-component systems. To eliminate this bottleneck, we present the polyply software suite that provides 1) a multi-scale graph matching algorithm designed to generate parameters quickly and for arbitrarily complex polymeric topologies, and 2) a generic multi-scale random walk protocol capable of setting up complex systems efficiently and independent of the target force-field or model resolution. We benchmark quality and performance of the approach by creating realistic coordinates for polymer melt simulations, single-stranded as well as circular single-stranded DNA. We further demonstrate the power of our approach by setting up a microphase-separated block copolymer system, and by generating a liquid-liquid phase separated system inside a lipid vesicle.},
+	language = {en},
+	number = {1},
+	urldate = {2025-05-08},
+	journal = {Nature Communications},
+	author = {Grünewald, Fabian and Alessandri, Riccardo and Kroon, Peter C. and Monticelli, Luca and Souza, Paulo C. T. and Marrink, Siewert J.},
+	month = jan,
+	year = {2022},
+	note = {Publisher: Nature Publishing Group},
+	keywords = {Computational biophysics, Polymer chemistry, Software, Theory and computation},
+	pages = {68},
+	file = {Full Text PDF:/Users/hannahturney/Zotero/storage/MJJLMPKU/Grünewald et al. - 2022 - Polyply\; a python suite for facilitating simulations of macromolecules and nanomaterials.pdf:application/pdf},
+}
+
+@article{hayashi_radonpy_2022,
+	title = {{RadonPy}: automated physical property calculation using all-atom classical molecular dynamics simulations for polymer informatics},
+	volume = {8},
+	copyright = {2022 The Author(s)},
+	issn = {2057-3960},
+	shorttitle = {{RadonPy}},
+	url = {https://www.nature.com/articles/s41524-022-00906-4},
+	doi = {10.1038/s41524-022-00906-4},
[…]
