### 2025-05-08 8b270d9
addressing reviewer comments in issue #2
