+++ paper/paper.md
@@ -34,7 +34,7 @@ Polymer MD studies showcase an assortment of approaches to manually incorporate 
 
 Open-source software packages designed to build *in silico* polymer chains are focused on the design of polymers at the monomer and single-chain scale [@davel_parameterization_2024; @klein_hierarchical_2016; @santana-bonilla_modular_2023]. 
 
-Packages that have the capability to build multi-chain systems such as Polyply [@grunewald_polyply_2022], RadonPy [@hayashi_radonpy_2022], Polymer Structure Predictor [sahu_polymer_2022], do not contain in-built functions to incorporate molecular weight diversity, and put the onus on the user to generate an ensemble polydisperse chain sequences prior to using their chain building functions. Polymatic [@abbott_polymatic_2013], an algorithm to simulate polymerization, can theoretically create polydisperse systems through its random bond formation scheme, but does not have to ability to simultaneously control other key attributes of polymers such as blockiness and monomer ratio. The python tool Hoobas [@girard_hoobas_2019] builds polydisperse systems of randomized polymer chains for coarse-grained models. However, there is not currently a software package available that contains in-built functions to integrate multiple smaller-scale characteristics (monomer ratio, blockiness, degree of polymerization, chain terminal) into \textbf{atomistic} computational polymer models, whilst effectively capturing the heterogeneity and polydispersity of real-life samples. The development of SwiftPol was driven by the need to fill this gap in multi-scale building functionality of existing polymer building packages, to enable the simulation of realistic polymer models whilst allowing close user control of multiple system characteristics.
+Packages that have the capability to build multi-chain systems such as Polyply [@grunewald_polyply_2022], RadonPy [@hayashi_radonpy_2022], and Polymer Structure Predictor [sahu_polymer_2022], do not contain in-built functions to incorporate molecular weight diversity, and put the onus on the user to generate an ensemble of polydisperse chain sequences prior to using their chain building functions. Polymatic [@abbott_polymatic_2013], an algorithm to simulate polymerization, can theoretically create polydisperse systems through its random bond formation scheme, but does not have to ability to simultaneously control other key attributes of polymers such as blockiness and monomer ratio. The python tool Hoobas [@girard_hoobas_2019] builds polydisperse systems of randomized polymer chains for coarse-grained models. However, there is not currently a software package available that contains in-built functions to integrate multiple smaller-scale characteristics (monomer ratio, blockiness, degree of polymerization, chain terminal) into \textbf{atomistic} computational polymer models, whilst effectively capturing the heterogeneity and polydispersity of real-life samples. The development of SwiftPol was driven by the need to fill this gap in multi-scale building functionality of existing polymer building packages, to enable the simulation of realistic polymer models whilst allowing close user control of multiple system characteristics.
 
 
 SwiftPol uses open-source Python libraries RDkit [@landrum_rdkitrdkit_2024], OpenFF-interchange [@thompson_openff_2024], and OpenFF-toolkit [@wagner_openforcefieldopenff-toolkit_2024] to promote reproducibility and portability.  We have ensured that SwiftPol objects can be seamlessly integrated into existing open-source software built for parameterization and simulation, to allow the user to select their preferred force field, topology format, and engine. RDkit, OpenFF-interchange and OpenFF-toolkit enable the export of SwiftPol polymer ensembles directly to simulation engines, and to a range of MD-compatible file formats, including .pdb, .top, .prmtop, and .json.
