### 2025-05-08 matta-research-group/SwiftPol#22 @hannaomi
Hi Riccardo,

Apologies for the delay in addressing this comment. These papers (and another package - Polymatic) are discussed in the newest version of the paper. New proof available for download [here](https://github.com/matta-research-group/SwiftPol/actions/runs/14905561897/artifacts/3085750319).
