### 2025-05-08 a3f6ffa
Better way of generating chain lengths
