### 2025-05-08 4d26e74
remove redundancy in build tests
