### 2025-05-08 c679f2c
remove redundancy in build tests
