### 2025-05-08 8b5fe19
Add JOSS banner to README
