### 2025-05-08 2622829
Remove depracated solavtion test
