### 2025-05-08 #8053 @ricalessandri
With this last point addressed (https://github.com/matta-research-group/SwiftPol/issues/22), my checklist is done, the authors addressed all my comments. @srmnitc please let me know if anything else is needed from my side.
