### 2025-05-16 matta-research-group/SwiftPol#23 @hmacdope
Demo notebook for going all the way to simulation
While the demo notebooks show some fantastic structures and the power of the approach, as far as I could see there was no demo of how to actually RUN a simulation from an engine, e.g with OpenMM.

This would massively help users take the next step
