### 2025-05-16 matta-research-group/SwiftPol#24 @hmacdope
Remove lint
Nitpick, but its better to remove lint like

- .vscode
- .DS_Store

These can easily be added to .gitignore if they creep back in.
