### 2025-05-16 matta-research-group/SwiftPol#25 @hmacdope
clarify /build/lib/swiftpol vs /swiftpol
There seem to be two directories with core functionality in them

 /build/lib/swiftpol vs /swiftpol at the root level. Would be nice to clean these up.
