### 2025-05-16 #8053 @hmacdope
Some initial notes on @hannaomi on what is some very cool (and useful) work.

- [ ] RDkit -> RDKit in paper
- [ ] https://github.com/matta-research-group/SwiftPol/issues/23 addressing this will help users put this into production
- [ ] Visulisation and or a short simulation + prelim analysis  of Product X system would be of value I think to show applicability for MD directly.
- [ ] Speed benchmarking -> Performance benchmarking
- [ ] https://github.com/matta-research-group/SwiftPol/issues/24
- [ ] https://github.com/matta-research-group/SwiftPol/issues/25
