### 2025-05-17 70c0ffd
Simulation in PLGA_demo notebook to address issue #23
