### 2025-05-17 ab56429
Removed lint and added files to .gitignore to address issue #24
