### 2025-05-17 6a6e1ca
removed redundant code files - issue #25
