### 2025-05-17 #8053 @hannaomi
Addressed changes requested by @hmacdope in [paper branch commits](https://github.com/matta-research-group/SwiftPol/commits/paper/) on May 17th 2025.
