### 2025-05-23 7029d33
Added unique atom name generation for residuals
