### 2025-05-23 d8c06ca
Removed lint and added to git ignore
