### 2025-05-23 9fefa37
Add A_target attribute for homopolymers
