### 2025-05-23 639e510
debug residual calculation for homopolymers
