### 2025-05-23 922b192
debug residual calculation for homopolymers
