### 2025-05-23 270fe20
debug residual calculation for homopolymers
