### 2025-05-23 8e3acf4
debug residual calculation for homopolymer
