### 2025-05-30 matta-research-group/SwiftPol#23 @jfaraudo
I have a problem with PLGA_demo notebook when calling packmol to add water to the structure.
