### 2025-05-30 matta-research-group/SwiftPol#23 @jfaraudo
CalledProcessError                        Traceback (most recent call last)
File ~/anaconda3/envs/swiftpol/lib/python3.11/site-packages/openff/interchange/components/_packmol.py:725, in pack_box(molecules, number_of_copies, solute, tolerance, box_vectors, target_density, box_shape, center_solute, working_directory, retain_working_files)
    724 try:
--> 725     result = subprocess.check_output(
    726         packmol_path,
    727         stdin=file_handle,
    728         stderr=subprocess.STDOUT,
    729     )
    730 except subprocess.CalledProcessError as error:
    731     # Custom error codes seem to start at 170
    732     # https://github.com/m3g/packmol/blob/v20.15.1/src/exit_codes.f90#L13-L16

File ~/anaconda3/envs/swiftpol/lib/python3.11/subprocess.py:466, in check_output(timeout, *popenargs, **kwargs)
    464     kwargs['input'] = empty
--> 466 return run(*popenargs, stdout=PIPE, timeout=timeout, check=True,
    467            **kwargs).stdout

File ~/anaconda3/envs/swiftpol/lib/python3.11/subprocess.py:571, in run(input, capture_output, timeout, check, *popenargs, **kwargs)
    570     if check and retcode:
--> 571         raise CalledProcessError(retcode, process.args,
    572                                  output=stdout, stderr=stderr)
    573 return CompletedProcess(process.args, retcode, stdout, stderr)

CalledProcessError: Command '/home/jordifaraudo/anaconda3/envs/swiftpol/bin/packmol' returned non-zero exit status 171.

PACKMOLRuntimeError                       Traceback (most recent call last)
Cell In[11], line 11
      4 molecules, number_of_copies, topology, box_vectors, residual_monomer_actual = build.calculate_box_components(chains = polymer.chains,
      5                                                                                                        monomers = monomer_smiles,
      6                                                                                                        sequence = polymer.sequence,
      7                                                                                                        salt_concentration= 1 * unit.mole/unit.litre,
      8                                                                                                        residual_monomer = 10,
      9                                                                                                        solvated=True)
     10 import openff.interchange.components._packmol as packmol_wrapper
---> 11 solvated_top = packmol_wrapper.pack_box(molecules = molecules + polymer.chains,
     12                                         number_of_copies = number_of_copies+[1 for i in range(len(polymer.chains))],
     13                                         box_vectors = box_vectors,
     14                                         tolerance = 1*unit.angstrom)
     15 solvated_top.visualize()

File ~/anaconda3/envs/swiftpol/lib/python3.11/site-packages/openff/utilities/utilities.py:81, in requires_package.<locals>.inner_decorator.<locals>.wrapper(*args, **kwargs)
     78 except Exception as e:
     79     raise e
---> 81 return function(*args, **kwargs)

File ~/anaconda3/envs/swiftpol/lib/python3.11/site-packages/openff/interchange/components/_packmol.py:735, in pack_box(molecules, number_of_copies, solute, tolerance, box_vectors, target_density, box_shape, center_solute, working_directory, retain_working_files)
    730     except subprocess.CalledProcessError as error:
    731         # Custom error codes seem to start at 170
    732         # https://github.com/m3g/packmol/blob/v20.15.1/src/exit_codes.f90#L13-L16
    733         open("packmol_error.log", "w").write(error.stdout.decode("utf-8"))
--> 735         raise PACKMOLRuntimeError(
    736             f"PACKMOL failed with error code {error.returncode}. Wrote file packmol_error.log in working "
    737             "directory, which might be a temporary directory. Set the argument `working_directory` to "
    738             "point this to a persistent path.",
    739         ) from error
    741     packmol_succeeded = result.decode("utf-8").find("Success!") > 0
    743 if not packmol_succeeded:

PACKMOLRuntimeError: PACKMOL failed with error code 171. Wrote file packmol_error.log in working directory, which might be a temporary directory. Set the argument `working_directory` to point this to a persistent path.
