### 2025-05-30 matta-research-group/SwiftPol#23 @jfaraudo
Also note that packmol is installed and apparently working:

(swiftpol) [jordifaraudo@pcd313c SwiftPol]$ which packmol
~/anaconda3/envs/swiftpol/bin/packmol
