### 2025-05-30 matta-research-group/SwiftPol#23 @hannaomi
Hi @jfaraudo ,

Glad to hear the install issues were resolved.
Please could you add the following arguments to the packmol_wrapper.packbox function:

```
retain_working_files=True,
working_directory = <desired directory here>
```
making the entire function call read:

```
solvated_top = packmol_wrapper.pack_box(molecules = molecules + polymer.chains,
                                        number_of_copies = number_of_copies+[1 for i in range(len(polymer.chains))],
                                        box_vectors = box_vectors,
                                        tolerance = 1*unit.angstrom,
                                        retain_working_files=True,
                                        working_directory = <desired directory here>)

```

this will generate a file named 'packmol_error.log'.

Please could you attach that output here so I can debug?

Many thanks!
