### 2025-05-30 matta-research-group/SwiftPol#23 @jfaraudo
Done! This is the output file:
################################################################################

 PACKMOL - Packing optimization for the automated generation of
 starting configurations for molecular dynamics simulations.

                                                             Version 20.16.1

################################################################################

  Packmol must be run with: packmol < inputfile.inp

  Userguide at: http://m3g.iqm.unicamp.br/packmol

  Reading input file... (Control-C aborts)
  Types of coordinate files specified: pdb
  ERROR: Number of molecules of type            1  set to less than 1
STOP 171
