### 2025-05-30 matta-research-group/SwiftPol#23 @jfaraudo
[packmol_output.zip](https://github.com/user-attachments/files/20524094/packmol_output.zip)
