### 2025-05-30 matta-research-group/SwiftPol#23 @hannaomi
My apologies that you have encountered this issue. The only other fix I can advise at this point is to pull the most recent commit and reinstall using `pip install .` , although it seems like this has been performed recently and the line of code that is causing this issue was pushed some time ago
