### 2025-05-30 matta-research-group/SwiftPol#23 @jfaraudo
I have executed the command you asked for, this is the result:

(swiftpol) [jordifaraudo@pcd313c SwiftPol]$ git rev-parse HEAD
5e8373a20ea498f0897c93f426eeecaa66989040
