### 2025-05-30 matta-research-group/SwiftPol#23 @jfaraudo
The install was made from main branch, using this:

git clone https://github.com/matta-research-group/SwiftPol

cd SwiftPol
# install requirements into new environment
conda env create --file Dev_tools/swiftpol.yml

conda activate swiftpol

pip install -e .
