### 2025-05-30 matta-research-group/SwiftPol#23 @jfaraudo
And yes, the solvated=True command was present.
Here is the cell:

#Use the OpenFF Packmol Wrapper to solvate the chains
#Example of using existing tools to prepare SwiftPol chains for Molecular Dynamics
from openff.units import unit
molecules, number_of_copies, topology, box_vectors, residual_monomer_actual = build.calculate_box_components(chains = polymer.chains,
                                                                                                       monomers = monomer_smiles,
                                                                                                       sequence = polymer.sequence,
                                                                                                       salt_concentration= 1 * unit.mole/unit.litre,
                                                                                                       residual_monomer = 10,
                                                                                                       solvated=True)
import openff.interchange.components._packmol as packmol_wrapper
solvated_top = packmol_wrapper.pack_box(molecules = molecules + polymer.chains,
                                        number_of_copies = number_of_copies+[1 for i in range(len(polymer.chains))],
                                        box_vectors = box_vectors,
                                        tolerance = 1*unit.angstrom,
                                        retain_working_files=True,
                                        working_directory="/home/jordifaraudo/SwiftPol/")
solvated_top.visualize()
