### 2025-05-30 matta-research-group/SwiftPol#23 @hannaomi
Thanks for this extra information. I will try and reproduce the error you're seeing
