### 2025-05-30 e25076e
Change calculate_box_components solvation calculation
