+++ tests/build_tests.py
@@ -228,13 +228,12 @@ class TestCalculateBoxComponents(unittest.TestCase):
         #print('resid_mon = ', residual_monomer_actual)
         #self.assertTrue(1.20 <= residual_monomer_actual <= 1.80)
 
-        # Calculate box components - test case without residual monomer and solvent
+        # Calculate box components - test case without residual monomer
         molecules, number_of_copies, topology, box_vectors, residual_monomer_actual = build.calculate_box_components(chains = x.chains, 
                                                                                                             monomers = x.monomers, 
                                                                                                             sequence = x.sequence, 
                                                                                                             salt_concentration= 0 * unit.mole/unit.litre,
-                                                                                                            residual_monomer = 0,
-                                                                                                            solvated=False      
+                                                                                                            residual_monomer = 0  
         )
         # Check if the returned molecules is a list of 5 objects, and water is not included
         self.assertTrue(isinstance(molecules, list))
