### 2025-05-30 5f8de95
Adjustment to tests to ensure components calculated correctly
