### 2025-05-30 88dccfe
adjustment to attempt to fix water calculation
