### 2025-05-30 af8f0e2
debugging in water calc - remove units
