### 2025-05-30 52c617a
remove units from build.py salt and water cal
