### 2025-05-30 2b25d2f
rework calculate_box_components
