### 2025-05-30 3e7ae08
amended tests to include units
