+++ tests/build_tests.py
@@ -322,7 +322,7 @@ class TestPolymerSystem(unittest.TestCase):
         #Test packmol packing
         y = build.polymer_system(monomer_list=['O[C@H](C)C(=O)O[I]'], 
                                 reaction = '[C:1][O:2][H:3].[I:4][O:5][C:6]>>[C:1][O:2][C:6].[H:3][O:5][I:4]',
-                                length_target=5,
+                                length_target=10,
                                 num_chains = 1,
                                 copolymer=False,
                                 acceptance=10)
