### 2025-06-01 402d7ed
Adjustment to components calculation to reflect main branch debug
