### 2025-06-01 matta-research-group/SwiftPol#23 @hannaomi
Thanks for your patience @jfaraudo .

I have debugged the build.calculate_box_components function. The PLGA_demo notebook should run for you now.

Please let me know if you run into any problems.
Safe travels!
