### 2025-06-01 matta-research-group/SwiftPol#23 @jfaraudo
Thanks! What was the origin of the problem? Something related to packmol input file?
