### 2025-06-01 matta-research-group/SwiftPol#23 @hannaomi
The calculation of the box size, and subsequent water molecule count was an issue for some smaller polymer systems and causing a water count of 0 to be exported to packmol, which leads to the packmol error.
