### 2025-06-02 matta-research-group/SwiftPol#23 @jfaraudo
This issue can be closed, I think.
