### 2025-06-04 13f626f
Adjusted residual oligomer to be polydisperse
