+++ swiftpol/build.py
@@ -344,8 +344,7 @@ def calculate_box_components(chains, monomers, sequence, salt_concentration = 0.
         The desired salt concentration in the simulation box. Defaults to 0 M.
     residual_monomer : float, optional
         The desired residual monomer concentration in the simulation box. Defaults to 0.00%.
-    solvated : bool, optional
-        Indicates whether the system contains water. Defaults to False.
+
     
     Returns:
     --------
@@ -521,7 +520,6 @@ def introduce_stereoisomers(stereo_monomer, instance, seq):
     modified_seq = ''.join(seq_list)
     return modified_seq
 
-#Class object for generic polymer system
 
 class polymer_system:
     try:
@@ -936,7 +934,7 @@ class polymer_system:
         print(f'Polyply input files generated! Saved at {return_tuple}')
         return return_tuple
 
-    def calculate_residuals(self, residual_monomer = 0, residual_oligomer = 0):
+    def calculate_residuals(self, residual_monomer = 0, residual_oligomer = 0, return_rdkit = False):
         """
         Generate residual monomer and oligomer molecules, and molecule counts.
 
@@ -946,6 +944,8 @@ class polymer_system:
             The desired residual monomer concentration in the simulation box. Default is 0.00.
         residual_oligomer : float
             The desired residual oligomer concentration in the simulation box. Default is 0.00.
+        return_rdkit : bool, optional
+            If True, return RDKit molecule objects instead of OpenFF Molecule objects. Default is False.
             
         Returns
         -------
@@ -981,7 +981,7 @@ class polymer_system:
         for m in topology.molecules:
             rolling_mass += sum(atom.mass for atom in m.atoms)
         if rolling_mass.magnitude < 1000:
-            warnings.warn('Residual monomer/oligomer calculation may not be accurate for small systems, as the residual mass can be close to or lower than the weight of a single monomer. Please check the output by querying the value of residual_monomer_actual')
+            warnings.warn('Residual monomer/oligomer calculation may not be appropriate for small systems, as the residual mass can be close to or lower than the weight of a single monomer. Please check the output by querying the value of residual_monomer_actual')
         number_of_copies = []
         # residual to add
         monomer_to_add = (rolling_mass.magnitude/100 - residual_monomer - residual_oligomer) * residual_monomer
@@ -1005,6 +1005,7 @@ class polymer_system:
             B = Molecule.from_rdkit(B_rd)
             B_mass = sum([atom.mass for atom in B.atoms])
             B.name = 'B_residual'
+            mon_count = 0
             for r in range(1,100):
                 if (r * A_mass.magnitude) + (r * B_mass.magnitude) >= monomer_to_add:
                     mon_count = r
@@ -1024,6 +1025,7 @@ class polymer_system:
             A.name = 'A_residual'
             B = Molecule.from_smiles('C')
             B.name = 'B_residual'
+            mon_count = 0
             for r in range(1,100):
                 if (r * A_mass.magnitude) >= monomer_to_add:
                     mon_count = r
@@ -1036,7 +1038,11 @@ class polymer_system:
         oligomers = []
         if 'A' in sequence and 'B' in sequence:
             for i in range(1000):
-                oligo_seq = reduce(lambda x, y: x + y, np.random.choice(['A', 'B'], size=(int(self.length_target * 0.1)), p=[self.A_target/100,1-(self.A_target/100)]))
+                sigma = np.sqrt(np.log(1.05))  
+                mu = np.log(self.length_target * 0.1) - 0.5 * sigma**2  
+                chain_lengths = np.random.lognormal(mu, sigma, size=1)
+                chain_lengths = np.round(chain_lengths).astype(int)  
[…]
