### 2025-06-09 66f9712
Testing new approach for copolymer building from PDI value
