### 2025-06-09 56a2ca3
adjusted tests to speed up
