### 2025-06-09 72f609e
adjusted tests to speed up
