### 2025-06-10 9d6ef53
adjusted tests to speed up
