### 2025-06-11 b756db3
Enable SMILES terminals in build.build_polymer
