### 2025-06-11 3e4093d
hydrogen terminals as standard for build_system
