### 2025-06-15 #8053 @hannaomi
Hi @srmnitc ,

I can confirm I have double checked the authors and affiliations incl. ORCIDs.
The version number from the zenodo release is v.0.1.3.
DOI is https://doi.org/10.5281/zenodo.15667457

license and author list match
