### 2025-06-16 matta-research-group/SwiftPol#37 @srmnitc
Small changes for JOSS paper
This is part of the review at https://github.com/openjournals/joss-reviews/issues/8053

Some small suggestions for the JOSS paper:

- line 58, 61 does SMILES and SMARTS need citations?
- line 146 - spaloma charge , package name needs a hyphen in between.
