### 2025-06-16 8c7fa30
Fix issue #37 for JOSS review
