+++ paper/paper.bib
@@ -354,6 +354,24 @@ Fix misnamed test
 	file = {1-s2.0-S0378517321007134-main.pdf:/Users/hannahturney/Zotero/storage/KM3E4T6L/1-s2.0-S0378517321007134-main.pdf:application/pdf;Wan et al. - 2021 - Effect of polymer source on in vitro drug release .pdf:/Users/hannahturney/Zotero/storage/X6H6UJXY/Wan et al. - 2021 - Effect of polymer source on in vitro drug release .pdf:application/pdf},
 }
 
+@misc{daylight_chemical_information_systems_inc_daylight_2022,
+	title = {Daylight {Theory}: {SMARTS} - {A} {Language} for {Describing} {Molecular} {Patterns}},
+	url = {https://daylight.com/dayhtml/doc/theory/theory.smarts.html},
+	urldate = {2024-08-29},
+	author = {Daylight Chemical Information Systems, Inc.},
+	year = {2022},
+	file = {Daylight Theory\: SMARTS - A Language for Describing Molecular Patterns:/Users/hannahturney/Zotero/storage/32CI89JX/theory.smarts.html:text/html},
+}
+
+@misc{daylight_chemical_information_systems_inc_daylight_2022-1,
+	title = {Daylight {Theory}: {SMILES} - {A} {Simplified} {Chemical} {Language}},
+	url = {https://www.daylight.com/dayhtml/doc/theory/theory.smiles.html},
+	urldate = {2024-08-29},
+	author = {Daylight Chemical Information Systems, Inc.},
+	year = {2022},
+	file = {Daylight Theory\: SMILES:/Users/hannahturney/Zotero/storage/LFZAJF4Y/theory.smiles.html:text/html},
+}
+
 @article{ochi_influence_2021,
 	title = {Influence of {PLGA} molecular weight distribution on leuprolide release from microspheres},
 	volume = {599},
@@ -475,7 +493,7 @@ Fix misnamed test
 	month = jan,
 	year = {2022},
 	note = {Publisher: Nature Publishing Group},
-	keywords = {Computational biophysics, Polymer chemistry, Software, Theory and computation},
+	keywords = {Theory and computation, Computational biophysics, Polymer chemistry, Software},
 	pages = {68},
 	file = {Full Text PDF:/Users/hannahturney/Zotero/storage/MJJLMPKU/Grünewald et al. - 2022 - Polyply\; a python suite for facilitating simulations of macromolecules and nanomaterials.pdf:application/pdf},
 }
@@ -497,7 +515,7 @@ Fix misnamed test
 	month = nov,
 	year = {2022},
 	note = {Publisher: Nature Publishing Group},
-	keywords = {Atomistic models, Computational methods, Polymers},
+	keywords = {Polymers, Atomistic models, Computational methods},
 	pages = {1--15},
 	file = {Full Text PDF:/Users/hannahturney/Zotero/storage/79N6Q3NX/Hayashi et al. - 2022 - RadonPy automated physical property calculation using all-atom classical molecular dynamics simulat.pdf:application/pdf},
 }
@@ -554,7 +572,7 @@ Fix misnamed test
 	author = {Abbott, Lauren J. and Hart, Kyle E. and Colina, Coray M.},
 	month = jan,
 	year = {2013},
-	keywords = {Equilibration, Glassy polymers, Molecular simulations, Polymers, Simulated polymerization},
+	keywords = {Polymers, Equilibration, Glassy polymers, Molecular simulations, Simulated polymerization},
 	pages = {1334},
 	file = {Full Text PDF:/Users/hannahturney/Zotero/storage/UQWZW8P8/Abbott et al. - 2013 - Polymatic a generalized simulated polymerization algorithm for amorphous polymers.pdf:application/pdf},
 }
