### 2025-06-17 39c1319
Fixed reference compilation issue for JOSS review
