+++ paper/paper.bib
@@ -363,15 +363,6 @@ Fix misnamed test
 	file = {Daylight Theory\: SMARTS - A Language for Describing Molecular Patterns:/Users/hannahturney/Zotero/storage/32CI89JX/theory.smarts.html:text/html},
 }
 
-@misc{daylight_chemical_information_systems_inc_daylight_2022-1,
-	title = {Daylight {Theory}: {SMILES} - {A} {Simplified} {Chemical} {Language}},
-	url = {https://www.daylight.com/dayhtml/doc/theory/theory.smiles.html},
-	urldate = {2024-08-29},
-	author = {Daylight Chemical Information Systems, Inc.},
-	year = {2022},
-	file = {Daylight Theory\: SMILES:/Users/hannahturney/Zotero/storage/LFZAJF4Y/theory.smiles.html:text/html},
-}
-
 @article{ochi_influence_2021,
 	title = {Influence of {PLGA} molecular weight distribution on leuprolide release from microspheres},
 	volume = {599},
@@ -575,4 +566,13 @@ Fix misnamed test
 	keywords = {Polymers, Equilibration, Glassy polymers, Molecular simulations, Simulated polymerization},
 	pages = {1334},
 	file = {Full Text PDF:/Users/hannahturney/Zotero/storage/UQWZW8P8/Abbott et al. - 2013 - Polymatic a generalized simulated polymerization algorithm for amorphous polymers.pdf:application/pdf},
+}
+
+@misc{daylight_chemical_information_systems_inc_daylight_2022-1,
+	title = {Daylight {Theory}: {SMILES}},
+	url = {https://www.daylight.com/dayhtml/doc/theory/theory.smiles.html},
+	urldate = {2025-06-17},
+	author = {{Daylight Chemical Information Systems, Inc.}},
+	year = {2022},
+	file = {Daylight Theory\: SMILES:/Users/hannahturney/Zotero/storage/PZ4LJ44Y/theory.smiles.html:text/html},
 }
