### 2025-06-23 #8053 @xuanxu
I'd say that the label for the tables (`Table 1` and `Table 2`) must be in the caption, see https://joss.readthedocs.io/en/latest/paper.html#tables-and-figures
