### 2025-06-23 14d3aa2
fixed paper table labelling
