+++ paper/paper.md
@@ -70,7 +70,7 @@ The input values for the SwiftPol builder, seen in \autoref{tab:Table 1}, were t
 \begin{flushleft}
 \begin{table}[h!]
 \captionsetup{justification=raggedright,singlelinecheck=false}
-\caption[Input parameters for SwiftPol PLGA builder function, for the building of product X.\label{Table 1}]
+\caption[Input parameters for SwiftPol PLGA builder function, for the building of product X.\label{tab:Table 1}]
 \begin{tabular}{|l|l|}
 \hline
 \textbf{INPUT} & \textbf{VALUE} \\
@@ -93,7 +93,7 @@ NACL CONCENTRATION (M) & 0.1 \\
 \begin{flushleft}
 \begin{table}[h!]
 \captionsetup{justification=raggedright,singlelinecheck=false}
-\caption[SwiftPol system build attributes. \( \bar{x}_n \) = mean value of attribute across n chains.\label{Table 2}]
+\caption[SwiftPol system build attributes. \( \bar{x}_n \) = mean value of attribute across n chains.\label{tab:Table 2}]
 \begin{tabular}{|l|l|}
 \hline
 \textbf{ATTRIBUTE} & \textbf{\( \bar{x}_n \)} \\
