### 2025-06-23 50391e3
fix in text referencing of tables
