### 2025-06-23 36173df
moved labels to same line as captions
