### 2025-06-23 #8053 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
