+++ paper/paper.md
@@ -61,7 +61,7 @@ SwiftPol also contains functions to generate conformers using RDKit [@landrum_rd
 # Application: building a poly(lactide-co-glycolide) ensemble
 
 Using SwiftPol, we have successfully constructed polydisperse ensembles of poly(lactide-co-glycolide) (PLGA), a widely used biodegradable polymer. We used the molecular structures and properties of experimental PLGA products as input for SwiftPol building functions to create representative PLGA systems to be used for molecular dynamics simulations. By integrating experimental data, such as chain terminals, copolymer ratios of lactic and glycolic acid, and blockiness, we have been able to replicate the bulk characteristics of various commercial polymer products, namely polydispersity. 
-A full example implementation of SwiftPol for building PLGA systems can be found in the [building a PLGA system example notebook.](Example_Notebooks/PLGA_demo.ipynb)
+A full example implementation of SwiftPol for building PLGA systems can be found in the [building a PLGA system example notebook.](https://github.com/matta-research-group/SwiftPol/blob/main/Example_Notebooks/PLGA_demo.ipynb)
 We used SwiftPol to build ‘product X’, a commercially available 75:25 LA:GA ester-terminated PLGA. Following the chain build, another SwiftPol function was used to calculate the appropriate box size for the unit cell, number of water molecules, NaCl ions, and residual monomer molecules to include in the simulation of a complete condensed polymer ensemble.
 The input values for the SwiftPol builder, seen in \autoref{tab:Table1}, were taken from quality assurance documents provided by the manufacturer of product X, except the value for blockiness which was measured experimentally by Sun et al [@sun_characterization_2022]. The system attributes assigned by SwiftPol to the completed condensed PLGA unit cell are in seen in \autoref{tab:Table2}.
 
