### 2025-06-23 #8053 @hannaomi
I have fixed a broken link in the text but otherwise the proof looks fine. Thanks all!
