### 2025-06-26 10dcd90
align versioning with github release
