### 2025-06-26 a278c5c
implement scm tools versioning
