### 2025-06-26 470b9ab
revert pyproject.toml
