### 2025-06-27 42bd54d
Remove italics notation
