### 2025-06-27 #8053 @hannaomi
I have removed the italics notation in the paper.md file. Please let me know if any other adjustments are needed to correct. Thanks all
