### 2025-06-29 deb7f56
Added citation info
