### 2025-06-29 63f5889
debugging errors in system build causing chain lengths <1
